[0.0.6] Non-Text content is now stored as attachments
In the content object, there is now an array called attachments that
stores each attachement struct with its data. This means that the user
will be able to access attachment properties such as type, subtype,
encoding, size, etc. directly on the attachment. When they want to do
something with the attachment data, they can access the attachment.data
property.