[0.0.9] Child folders now correctly serialize to JSON
Previously, child folders would cause an error on JSON serialization
because they maintain a property called 'parent'. This property is a
reference to the parent folder Object. When attempting to serialize this
property, a cyclic dependency error occurs and the entire result
serialization fails. To fix this, we recursively delete the parent
property from all folders and their children. Since the child folders
are will maintain structure (Inside the children property of their
parent folder), this does not remove any functionality.